requirements.txt
pandas==2.2.2
plotly-express==0.4.1
streamlit==1.39.0
matplotlib==3.10.6
